refactor(③ react): relocate react/supply drive loop onto impl VM (Stage 2 PR1)
Pure relocation, no behavior change. Moves the react/supply drive-loop
cluster (`run_react_event_loop`, `run_react_event_loop_drain`,
`run_react_consumer`, `drive_react_subscriptions`, `replay_static_supply`,
`run_react_close_callbacks` + the `extract_*` / `resolve_supply_channel_id`
helpers) from `impl Interpreter` (runtime/subtest.rs) onto `impl VM`
(new vm/vm_react_loop.rs).

Why: Stage 2 of the Track C react/supply VM-native work needs the
`whenever`-body dispatch to run compiled bytecode (`vm_call_on_value`)
instead of the tree-walking `call_sub_value`. The borrow checker forces
this: the VM owns the `Interpreter` by value (`VM::new` moves it), so a
`&mut Interpreter` method cannot construct a VM — the loop itself must
live on `impl VM`. This PR performs only the move; dispatch still goes
through `self.interpreter.call_sub_value` (valid because the VM owns the
interpreter), isolating the mechanical relocation from the upcoming
dispatch-semantics change.

Entry points:
- `react { ... }` (`exec_react_scope_op`, already a VM method) calls the
  relocated `self.run_react_event_loop()` directly.
- `await $supply` / `$supply.Promise` (`supply_promise_on_demand`, an
  Interpreter method) reaches the relocated loop through a thin
  `Interpreter::drive_react_subscriptions` bridge using the established
  `mem::take` / `VM::new` / `into_interpreter` dance (same pattern the
  native map loop uses).

Visibility: the supplier-registry accessors the loop polls
(`supplier_snapshot`, `next_supplier_id`, `supplier_register_promise`,
`take_promise_combinator_sources`) are now `pub(crate)` — they are
process-global registry functions, reachable from the VM by design.
`value_array_items`, `runtime_error_from_supply_reason`,
`call_supply_quit_handler` bumped to `pub(crate)`; `mod subtest` to
`pub(crate)`.

Verified: cargo build + clippy -D warnings clean; 40 t/ react/supply
tests; 64 whitelisted S17-supply/promise files (872 tests); make test
(7129); react + await output matches rakudo.
