+++ src/runtime/native_methods/mod.rs
@@ -23,16 +23,20 @@ pub(crate) use state::{
 };
 pub(crate) use state_lock::{acquire_lock, current_thread_id, lock_runtime_by_id, release_lock};
 
-// Re-export pub(in crate::runtime) items accessed from sibling `runtime` modules
+// Re-export items accessed from sibling `runtime` modules. A handful are also
+// re-exported `pub(crate)` below for the VM-side react/supply drive loop
+// (`crate::vm::vm_react_loop`): the supplier registry accessors it polls.
 pub(in crate::runtime) use state::{
     UdpBoundSocketState, allocate_async_listen_port, get_supply_collected_output, get_supply_taps,
-    lookup_async_listener, next_async_socket_id, next_supplier_id, next_supply_id, proc_stdin_map,
+    lookup_async_listener, next_async_socket_id, next_supply_id, proc_stdin_map,
     register_async_connection, register_promise_combinator_sources, register_supply_tap,
     register_udp_bound_socket, set_supply_collected_output, supplier_done, supplier_done_deferred,
-    supplier_emit, supplier_id_from_attrs, supplier_quit, supplier_register_promise,
-    supplier_reset, supplier_reset_keep_quit, supplier_snapshot, supply_channel_map,
-    supply_channel_map_pub, take_promise_combinator_sources, udp_port_in_use,
-    update_async_connection,
+    supplier_emit, supplier_id_from_attrs, supplier_quit, supplier_reset, supplier_reset_keep_quit,
+    supply_channel_map, supply_channel_map_pub, udp_port_in_use, update_async_connection,
+};
+// Supplier registry accessors driven by the VM-side react/supply loop.
+pub(crate) use state::{
+    next_supplier_id, supplier_register_promise, supplier_snapshot, take_promise_combinator_sources,
 };
 pub(in crate::runtime) use state_lock::next_lock_id;
 pub(in crate::runtime) use state_lock::next_semaphore_id;
