+++ src/vm.rs
@@ -55,6 +55,7 @@ mod vm_native_map;
 mod vm_native_sort;
 mod vm_native_subst;
 mod vm_native_test;
+mod vm_react_loop;
 mod vm_register_ops;
 mod vm_set_ops;
 pub(crate) mod vm_smart_match;
